requirements.txt
fastapi[standard]==0.116.1
uvicorn
httpx
motor==3.7.1
pydantic-settings==2.9.1
loguru
google-adk==1.8.0
